Add files via upload
I updated the database by adding several history tables so that when a record is deleted, the system will automatically save it and its related records to the corresponding history tables. 

For example: In the database, when a batch is delete, logically the ingredients used to produce this batch should also be deleted from the database, and the same with sold items which come from the deleted batch.