Add files via upload
When you first open up the database in MS Access, a login screen will pop up.
Here is the login credentials for your account:
  
   Username: xw236
   Password: Ilovewjr0811

Once logged in, you should be able to use the system.

I have inputted the following raw data into the system:

    Products: Oat Bran(Product ID:1); Other(Product ID: 2)
    Raw Materials: Whole Oat (Material ID: 1); Baking Soda(Material ID:2); Food Contact Bags( Material ID: 3)

In order to fully use the system, you should first input data in logical order:

    Add Suppliers: login, go to supply management, go to supplier management, click on "Add button", fill in the information of the supplier and save.
    Add purchasing orders: After finishing adding suppliers, you can go to supply management, then to purchasing order management, click on the "Add" button, fill in the information and save. 
    Once finishing with these two above, you can then go to production management and add batches. When adding a batch to the system, it is required to input the ingredients (raw material and its associated purchasing order ID). Besides, please note that when a new batch is added to the system, it will be automatically treated as unpacked and the quantity will be zero. So, if the added batch is already packed, you should edit the batch quantity by clicking the Batch Number on the Production Management page.
    After completing steps above, you can then add customers. Go to Sales Management, then to Customer Management, click the "Add" button, fill in the information and save.
    After adding customers, you can then add sales orders in the Sales Management. When adding a sales order, it is also required to input the order items along with their associated batches.

Also, it is important to keep in mind that whenever you are buying raw materials from a new supplier, you should add the supplier to the system first and then add the purchasing order. The same applies to the sales part. Whenever you are selling to a new customer, add the customer first, then the order.